fix: refactor posts
- Use shared `<Post />` component
- Infinately nested replies (WIP)
- Lighter topic badge colour
